pyproject.toml
[project]
name = "sistema-evento"
version = "0.1.0"
description = ""
authors = [
    {name = "Beatriz F"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
